+++ CHANGELOG.md
@@ -26,6 +26,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Require lazy type loader to return `Type` directly without an intermediary callable
 - Rename `ServerConfig` option `persistentQueryLoader` to `persistedQueryLoader`
 - Call previously unused methods `EnumType::parseValue()` and `EnumType::parseLiteral()`
+- Strongly type `PromiseAdapter::createRejected()` to require `\Throwable`
 
 ### Added